HPCC-20572 GraphsWidget + Open in new window
Opening the GraphsWidget in a new window loses its "mode" and thus fails to
display.
