+++ esp/src/eclwatch/templates/LFDetailsWidget.html
@@ -218,7 +218,7 @@
             </div>
             <div id="${id}_Queries" title="${i18n.Queries}" data-dojo-props="delayWidget: 'QuerySetQueryWidget', iconClass:'iconFind'" data-dojo-type="DelayLoadWidget">
             </div>
-            <div id="${id}_Graphs" title="${i18n.Graphs}" data-dojo-props="delayProps: {mode: 'LF'}, delayWidget: 'GraphsWidget', disabled: true" data-dojo-type="DelayLoadWidget">
+            <div id="${id}_Graphs" title="${i18n.Graphs}" data-dojo-props="delayWidget: 'GraphsLFWidget', disabled: true" data-dojo-type="DelayLoadWidget">
             </div>
             <div id="${id}_Workunit" title="${i18n.Workunit}" data-dojo-props="delayWidget: 'WUDetailsWidget'" data-dojo-type="DelayLoadWidget">
             </div>
