+++ esp/src/eclwatch/templates/QuerySetDetailsWidget.html
@@ -83,7 +83,7 @@
             </div>
             <div id="${id}_LibrariesUsed" title="${i18n.LibrariesUsed}" data-dojo-props="delayWidget: 'LibrariesUsedWidget'" data-dojo-type="DelayLoadWidget">
             </div>
-            <div id="${id}_Graphs" title="${i18n.Graphs}" data-dojo-props="delayProps: {mode: 'Query'}, delayWidget: 'GraphsWidget'" data-dojo-type="DelayLoadWidget">
+            <div id="${id}_Graphs" title="${i18n.Graphs}" data-dojo-props="delayWidget: 'GraphsQueryWidget'" data-dojo-type="DelayLoadWidget">
             </div>
             <div id="${id}_Resources" title="${i18n.Resources}" data-dojo-props="delayWidget: 'ResourcesWidget'" data-dojo-type="DelayLoadWidget">
             </div>
