+++ esp/src/eclwatch/templates/WUDetailsWidget.html
@@ -145,7 +145,7 @@
             </div>
             <div id="${id}_Timers" title="${i18n.Timers}" data-dojo-props="delayWidget: 'TimingPageWidget', disabled: true" data-dojo-type="DelayLoadWidget">
             </div>
-            <div id="${id}_Graphs" title="${i18n.Graphs}" data-dojo-props="delayProps: {mode: 'WU'}, delayWidget: 'GraphsWidget', disabled: true" data-dojo-type="DelayLoadWidget">
+            <div id="${id}_Graphs" title="${i18n.Graphs}" data-dojo-props="delayWidget: 'GraphsWUWidget', disabled: true" data-dojo-type="DelayLoadWidget">
             </div>
             <div id="${id}_Workflows" title="${i18n.Workflows}" data-dojo-props="delayWidget: 'WorkflowsWidget', disabled: true" data-dojo-type="DelayLoadWidget">
             </div>
